rename from Stripe to Payments provider
