+++ packages/core/src/types.ts
@@ -1,13 +1,14 @@
 import { CookieSerializeOptions } from 'cookie'
 import { Inngest } from 'inngest'
 import { type ChatCompletionRequestMessage } from 'openai-edge'
+import Stripe from 'stripe'
 import { z } from 'zod'
 
 import { CourseBuilderAdapter } from './adapters'
+import { CheckoutParams } from './lib/pricing/stripe-checkout'
 import { Cookie } from './lib/utils/cookie'
 import { LoggerInstance } from './lib/utils/logger'
 import { EmailListConfig, ProviderType, TranscriptionConfig } from './providers'
-import { StripeProviderConfig } from './providers/stripe'
 import { Coupon, MerchantCoupon, Price, Product, Purchase } from './schemas'
 import {
 	ContentResourceResourceSchema,
@@ -59,12 +60,67 @@ export interface RequestInternal {
 	error?: string
 }
 
+export interface PaymentsProviderConfig {
+	id: string
+	name: string
+	type: 'payment'
+	options: PaymentsProviderConsumerConfig
+	createCheckoutSession: (
+		checkoutParams: CheckoutParams,
+		adapter?: CourseBuilderAdapter,
+	) => Promise<{ redirect: string; status: number }>
+}
+
+export type PaymentsProviderConsumerConfig = Omit<
+	Partial<PaymentsProviderConfig>,
+	'options' | 'type'
+> & {
+	paymentsAdapter: PaymentsAdapter
+	errorRedirectUrl: string
+	cancelUrl: string
+	baseSuccessUrl: string
+}
+
+export interface PaymentsAdapter {
+	/**
+	 * Returns the percent off for a given coupon
+	 * @param identifier
+	 */
+	getCouponPercentOff(identifier: string): Promise<number>
+
+	/**
+	 * Returns a coupon id.
+	 *
+	 * TODO: these use the stripe types and we probably want to use an
+	 *   internal interface so that we can think about different providers
+	 *   in the future.
+	 * @param params
+	 */
+	createCoupon(params: Stripe.CouponCreateParams): Promise<string>
+
+	/**
+	 * Returns a promotion code.
+	 * @param params
+	 */
+	createPromotionCode(params: Stripe.PromotionCodeCreateParams): Promise<string>
+
+	/**
+	 * Returns the URL to redirect to for a checkout session.
+	 * @param params
+	 */
+	createCheckoutSession(
+		params: Stripe.Checkout.SessionCreateParams,
+	): Promise<string | null>
+
+	createCustomer(params: Stripe.CustomerCreateParams): Promise<string>
+}
+
 export type InternalProvider<T = ProviderType> = T extends 'transcription'
 	? TranscriptionConfig
 	: T extends 'email-list'
 		? EmailListConfig
 		: T extends 'checkout'
-			? StripeProviderConfig
+			? PaymentsProviderConfig
 			: never
 
 export interface InternalOptions<TProviderType = ProviderType> {
